Move common code to utils.js, cleanup.
