+++ client-src/elements/chromedash-guide-editall-page.js
@@ -1,6 +1,6 @@
 import {LitElement, css, html, nothing} from 'lit';
 import {ref} from 'lit/directives/ref.js';
-import {showToastMessage, flattenSections} from './utils.js';
+import {showToastMessage, flattenSections, setupScrollToHash} from './utils.js';
 import './chromedash-form-table';
 import './chromedash-form-field';
 import {
@@ -74,20 +74,12 @@ export class ChromedashGuideEditallPage extends LitElement {
     if (!el) return;
     await el.updateComplete;
 
-    window.addEventListener('beforeunload', (event) => {
-      // Having an event listener for beforeunload causes the browser to
-      // check for unsaved changes. But the event handler seems to be otherwise ignored.
-
-      event.preventDefault();
-      return event.returnValue = '';
-    });
-
     const hiddenTokenField = this.shadowRoot.querySelector('input[name=token]');
     hiddenTokenField.form.addEventListener('submit', (event) => {
       this.handleFormSubmit(event, hiddenTokenField);
     });
 
-    this.setupScrollToHash();
+    setupScrollToHash(this);
   }
 
   handleFormSubmit(event, hiddenTokenField) {
@@ -102,44 +94,6 @@ export class ChromedashGuideEditallPage extends LitElement {
 
   handleCancelClick() {
     window.location.href = `/guide/edit/${this.featureId}`;
-  }
-
-  setupScrollToHash() {
-    const scrollToElement = (hash) => {
-      if (hash) {
-        const el = this.shadowRoot.querySelector(hash);
-        if (el) {
-          // Find the header element for the form field.
-          const headerSelector = `chromedash-form-field[name="${el.name}"] tr th b`;
-          const header = this.shadowRoot.querySelector(headerSelector);
-          if (header) {
-            header.scrollIntoView({
-              block: 'start', inline: 'nearest', behavior: 'smooth',
-            });
-            // Focus on the corresponding form field.
-            if (el.input) {
-              // el.focus() calls el.input.focus();
-              el.focus();
-            } else {
-              // Not ready yet, so try after a timeout.  TODO: Avoid the timeout.
-              setTimeout(() => {
-                el.focus();
-              }, 0);
-            }
-          }
-        }
-      }
-    };
-
-    if (location.hash) {
-      const hash = decodeURIComponent(location.hash);
-      scrollToElement(hash);
-    }
-
-    // Add global function to jump to form field.
-    window.scrollToElement = (hash) => {
-      scrollToElement(hash);
-    };
   }
 
   renderSkeletons() {
